# Performance improvements using Struct.Equals rather than Struct Op_Equals