`LayoutObject` is called to many millions of times during the normal use of Anno Designer. In most setters there is a check for `_myPoint == default`. Following on from the below benchmarks, I have updated these getters to use `.Equals()` instead, which should have a small positive impact on performance.

|                  Method |     Mean |     Error |    StdDev |
|------------------------ |---------:|----------:|----------:|
|            Point_Equals | 5.046 ns | 0.1050 ns | 0.0982 ns |
|         Point_Op_Equals | 1.197 ns | 0.0372 ns | 0.0348 ns |
|    Point_Equals_Default | 4.848 ns | 0.0626 ns | 0.0555 ns |
| Point_Op_Equals_Default | 1.138 ns | 0.0172 ns | 0.0144 ns |

Benchmarks run on .Net 4.7.2, comparing 2 different Point structs from the WindowsBase.dll `System.Windows` namespace (the same Point implementation that is used in the code).

The *_Default benchmarks are testing `_myPoint == default` operations.