@Atria1234 2020-05-12
Just out of curiosity: what is the actual performance increase on that large layout?