@AgmasGold 2020-05-12
Its difficult to measure the performance exactly - might be a few points of a percent better? It definitely feels like a larger layout runs slightly faster with the changes though.